Uniformized task for QA with header
