+++ Common/TableProducer/PID/pidTOFFull.cxx
@@ -28,6 +28,7 @@
 #include "Common/DataModel/EventSelection.h"
 #include "TableHelper.h"
 #include "pidTOFBase.h"
+#include "qaTOF.h"
 #include "Framework/StaticFor.h"
 
 using namespace o2;
@@ -244,255 +245,11 @@ struct tofPidFull {
   PROCESS_SWITCH(tofPidFull, processNoEvTime, "Produce TOF response without TOF event time, standard for Run 2", true);
 };
 
-/// Task to produce the TOF QA plots
-struct tofPidFullQa {
-  static constexpr int Np = 9;
-  static constexpr const char* pT[Np] = {"e", "#mu", "#pi", "K", "p", "d", "t", "^{3}He", "#alpha"};
-  static constexpr std::string_view hexpected[Np] = {"expected/El", "expected/Mu", "expected/Pi",
-                                                     "expected/Ka", "expected/Pr", "expected/De",
-                                                     "expected/Tr", "expected/He", "expected/Al"};
-  static constexpr std::string_view hexpected_diff[Np] = {"expected_diff/El", "expected_diff/Mu", "expected_diff/Pi",
-                                                          "expected_diff/Ka", "expected_diff/Pr", "expected_diff/De",
-                                                          "expected_diff/Tr", "expected_diff/He", "expected_diff/Al"};
-  static constexpr std::string_view hexpsigma[Np] = {"expsigma/El", "expsigma/Mu", "expsigma/Pi",
-                                                     "expsigma/Ka", "expsigma/Pr", "expsigma/De",
-                                                     "expsigma/Tr", "expsigma/He", "expsigma/Al"};
-  static constexpr std::string_view hnsigma[Np] = {"nsigma/El", "nsigma/Mu", "nsigma/Pi",
-                                                   "nsigma/Ka", "nsigma/Pr", "nsigma/De",
-                                                   "nsigma/Tr", "nsigma/He", "nsigma/Al"};
-  static constexpr std::string_view hnsigmapt[Np] = {"nsigmapt/El", "nsigmapt/Mu", "nsigmapt/Pi",
-                                                     "nsigmapt/Ka", "nsigmapt/Pr", "nsigmapt/De",
-                                                     "nsigmapt/Tr", "nsigmapt/He", "nsigmapt/Al"};
-  static constexpr std::string_view hnsigmapospt[Np] = {"nsigmapospt/El", "nsigmapospt/Mu", "nsigmapospt/Pi",
-                                                        "nsigmapospt/Ka", "nsigmapospt/Pr", "nsigmapospt/De",
-                                                        "nsigmapospt/Tr", "nsigmapospt/He", "nsigmapospt/Al"};
-  static constexpr std::string_view hnsigmanegpt[Np] = {"nsigmanegpt/El", "nsigmanegpt/Mu", "nsigmanegpt/Pi",
-                                                        "nsigmanegpt/Ka", "nsigmanegpt/Pr", "nsigmanegpt/De",
-                                                        "nsigmanegpt/Tr", "nsigmanegpt/He", "nsigmanegpt/Al"};
-
-  HistogramRegistry histos{"Histos", {}, OutputObjHandlingPolicy::AnalysisObject};
-
-  Configurable<int> logAxis{"logAxis", 0, "Flag to use a log momentum axis"};
-  Configurable<int> nBinsP{"nBinsP", 400, "Number of bins for the momentum"};
-  Configurable<float> minP{"minP", 0.1f, "Minimum momentum in range"};
-  Configurable<float> maxP{"maxP", 5.f, "Maximum momentum in range"};
-  Configurable<int> nBinsDelta{"nBinsDelta", 200, "Number of bins for the Delta"};
-  Configurable<float> minDelta{"minDelta", -1000.f, "Minimum Delta in range"};
-  Configurable<float> maxDelta{"maxDelta", 1000.f, "Maximum Delta in range"};
-  Configurable<int> nBinsExpSigma{"nBinsExpSigma", 200, "Number of bins for the ExpSigma"};
-  Configurable<float> minExpSigma{"minExpSigma", 0.f, "Minimum ExpSigma in range"};
-  Configurable<float> maxExpSigma{"maxExpSigma", 200.f, "Maximum ExpSigma in range"};
-  Configurable<int> nBinsNSigma{"nBinsNSigma", 200, "Number of bins for the NSigma"};
-  Configurable<float> minNSigma{"minNSigma", -10.f, "Minimum NSigma in range"};
-  Configurable<float> maxNSigma{"maxNSigma", 10.f, "Maximum NSigma in range"};
-  Configurable<int> applyEvSel{"applyEvSel", 2, "Flag to apply rapidity cut: 0 -> no event selection, 1 -> Run 2 event selection, 2 -> Run 3 event selection"};
-  Configurable<bool> applyTrackCut{"applyTrackCut", false, "Flag to apply standard track cuts"};
-  Configurable<bool> applyRapidityCut{"applyRapidityCut", false, "Flag to apply rapidity cut"};
-
-  template <uint8_t i>
-  void addParticleHistos(const AxisSpec& pAxis, const AxisSpec& ptAxis)
-  {
-    // Exp signal
-    const AxisSpec expAxis{1000, 0, 2e6, Form("t_{exp}(%s) (ps)", pT[i])};
-    histos.add(hexpected[i].data(), "", kTH2F, {pAxis, expAxis});
-
-    // Signal - Expected signal
-    const AxisSpec deltaAxis{nBinsDelta, minDelta, maxDelta, Form("t-t_{ev}-t_{exp}(%s) (ps)", pT[i])};
-    histos.add(hexpected_diff[i].data(), "", kTH2F, {pAxis, deltaAxis});
-
-    // Exp Sigma
-    const AxisSpec expSigmaAxis{nBinsExpSigma, minExpSigma, maxExpSigma, Form("Exp_{#sigma}^{TOF}(%s) (ps)", pT[i])};
-    histos.add(hexpsigma[i].data(), "", kTH2F, {pAxis, expSigmaAxis});
-
-    // NSigma
-    const char* axisTitle = Form("N_{#sigma}^{TOF}(%s)", pT[i]);
-    const AxisSpec nSigmaAxis{nBinsNSigma, minNSigma, maxNSigma, axisTitle};
-    histos.add(hnsigma[i].data(), axisTitle, kTH2F, {pAxis, nSigmaAxis});
-    histos.add(hnsigmapt[i].data(), axisTitle, kTH2F, {ptAxis, nSigmaAxis});
-    histos.add(hnsigmapospt[i].data(), axisTitle, kTH2F, {ptAxis, nSigmaAxis});
-    histos.add(hnsigmanegpt[i].data(), axisTitle, kTH2F, {ptAxis, nSigmaAxis});
-  }
-
-  void init(o2::framework::InitContext&)
-  {
-    const AxisSpec multAxis{100, 0, 100, "TOF multiplicity"};
-    const AxisSpec vtxZAxis{100, -20, 20, "Vtx_{z} (cm)"};
-    const AxisSpec tofAxis{10000, 0, 2e6, "TOF Signal (ps)"};
-    const AxisSpec etaAxis{100, -1, 1, "#it{#eta}"};
-    const AxisSpec phiAxis{100, 0, TMath::TwoPi(), "#it{#phi}"};
-    const AxisSpec colTimeAxis{100, -2000, 2000, "Collision time (ps)"};
-    const AxisSpec colTimeResoAxis{100, 0, 1000, "#sigma_{Collision time} (ps)"};
-    const AxisSpec lAxis{100, 0, 500, "Track length (cm)"};
-    const AxisSpec ptResoAxis{100, 0, 0.1, "#sigma_{#it{p}_{T}}"};
-    AxisSpec ptAxis{nBinsP, minP, maxP, "#it{p}_{T} (GeV/#it{c})"};
-    AxisSpec pAxis{nBinsP, minP, maxP, "#it{p} (GeV/#it{c})"};
-    AxisSpec pExpAxis{nBinsP, minP, maxP, "#it{p}_{Exp. TOF} (GeV/#it{c})"};
-    if (logAxis) {
-      ptAxis.makeLogaritmic();
-      pAxis.makeLogaritmic();
-      pExpAxis.makeLogaritmic();
-    }
-
-    // Event properties
-    auto h = histos.add<TH1>("event/evsel", "", kTH1F, {{10, 0.5, 10.5, "Ev. Sel."}});
-    h->GetXaxis()->SetBinLabel(1, "Events read");
-    h->GetXaxis()->SetBinLabel(2, "Passed ev. sel.");
-    h->GetXaxis()->SetBinLabel(3, "Passed mult.");
-    h->GetXaxis()->SetBinLabel(4, "Passed vtx Z");
-
-    h = histos.add<TH1>("event/trackselection", "", kTH1F, {{10, 0, 10, "Selection passed"}});
-    h->GetXaxis()->SetBinLabel(1, "Tracks read");
-    h->GetXaxis()->SetBinLabel(2, "isGlobalTrack");
-    h->GetXaxis()->SetBinLabel(3, "hasITS");
-    h->GetXaxis()->SetBinLabel(4, "hasTOF");
-
-    histos.add("event/vertexz", "", kTH1F, {vtxZAxis});
-    h = histos.add<TH1>("event/particlehypo", "", kTH1F, {{10, 0, 10, "PID in tracking"}});
-    for (int i = 0; i < 9; i++) {
-      h->GetXaxis()->SetBinLabel(i + 1, PID::getName(i));
-    }
-    histos.add("event/trackmultiplicity", "", kTH1F, {multAxis});
-    histos.add("event/tofmultiplicity", "", kTH1F, {multAxis});
-    histos.add("event/colltime", "", kTH1F, {colTimeAxis});
-    histos.add("event/colltimereso", "", kTH2F, {multAxis, colTimeResoAxis});
-    histos.add("event/tofsignal", "", kTH2F, {pAxis, tofAxis});
-    histos.add("event/pexp", "", kTH2F, {pAxis, pExpAxis});
-    histos.add("event/eta", "", kTH1F, {etaAxis});
-    histos.add("event/phi", "", kTH1F, {phiAxis});
-    histos.add("event/etaphi", "", kTH2F, {etaAxis, phiAxis});
-    histos.add("event/length", "", kTH1F, {lAxis});
-    histos.add("event/pt", "", kTH1F, {ptAxis});
-    histos.add("event/p", "", kTH1F, {pAxis});
-    // histos.add("event/ptreso", "", kTH2F, {pAxis, ptResoAxis});
-
-    static_for<0, 8>([&](auto i) {
-      addParticleHistos<i>(pAxis, ptAxis);
-    });
-  }
-
-  template <o2::track::PID::ID id, typename T>
-  void fillParticleHistos(const T& t, const float& tof)
-  {
-    if (applyRapidityCut) {
-      const float y = TMath::ASinH(t.pt() / TMath::Sqrt(PID::getMass2(id) + t.pt() * t.pt()) * TMath::SinH(t.eta()));
-      if (abs(y) > 0.5) {
-        return;
-      }
-    }
-
-    const auto& nsigma = o2::aod::pidutils::tofNSigma<id>(t);
-    const auto& diff = o2::aod::pidutils::tofExpSignalDiff<id>(t);
-    histos.fill(HIST(hexpected[id]), t.p(), tof - diff);
-    histos.fill(HIST(hexpected_diff[id]), t.p(), diff);
-    histos.fill(HIST(hexpsigma[id]), t.p(), o2::aod::pidutils::tofExpSigma<id>(t));
-    histos.fill(HIST(hnsigma[id]), t.p(), nsigma);
-    histos.fill(HIST(hnsigmapt[id]), t.pt(), nsigma);
-    if (t.sign() > 0) {
-      histos.fill(HIST(hnsigmapospt[id]), t.pt(), nsigma);
-    } else {
-      histos.fill(HIST(hnsigmanegpt[id]), t.pt(), nsigma);
-    }
-  }
-
-  using Trks = soa::Join<aod::Tracks, aod::TracksExtra,
-                         aod::pidTOFFullEl, aod::pidTOFFullMu, aod::pidTOFFullPi,
-                         aod::pidTOFFullKa, aod::pidTOFFullPr, aod::pidTOFFullDe,
-                         aod::pidTOFFullTr, aod::pidTOFFullHe, aod::pidTOFFullAl,
-                         aod::TOFSignal, aod::TrackSelection>;
-  void process(soa::Join<aod::Collisions, aod::EvSels>::iterator const& collision,
-               Trks const& tracks)
-  {
-
-    histos.fill(HIST("event/evsel"), 1);
-    if (applyEvSel == 1) {
-      if (!collision.sel7()) {
-        return;
-      }
-    } else if (applyEvSel == 2) {
-      if (!collision.sel8()) {
-        return;
-      }
-    }
-
-    histos.fill(HIST("event/evsel"), 2);
-
-    // Computing Multiplicity first
-    float ntracks = 0;
-    int tofmult = 0;
-    for (auto t : tracks) {
-      if (applyTrackCut && !t.isGlobalTrack()) {
-        continue;
-      }
-      ntracks += 1;
-      if (!t.hasTOF()) { // Skipping tracks without TOF
-        continue;
-      }
-      tofmult++;
-    }
-    // if (0 && ntracks < 1) {
-    //   return;
-    // }
-    // if (0 && tofmult < 1) {
-    //   return;
-    // }
-    histos.fill(HIST("event/evsel"), 3);
-    if (abs(collision.posZ()) > 10.f) {
-      return;
-    }
-    histos.fill(HIST("event/evsel"), 4);
-    histos.fill(HIST("event/vertexz"), collision.posZ());
-    histos.fill(HIST("event/trackmultiplicity"), ntracks);
-    histos.fill(HIST("event/tofmultiplicity"), tofmult);
-
-    const float collisionTime_ps = collision.collisionTime() * 1000.f;
-    histos.fill(HIST("event/colltime"), collisionTime_ps);
-    histos.fill(HIST("event/colltimereso"), tofmult, collision.collisionTimeRes() * 1000.f);
-
-    for (auto t : tracks) {
-      histos.fill(HIST("event/trackselection"), 0.5f);
-      if (!t.isGlobalTrack()) { // Skipping non global tracks
-        continue;
-      }
-      histos.fill(HIST("event/trackselection"), 1.5f);
-      if (!t.hasITS()) { // Skipping tracks without ITS
-        continue;
-      }
-      histos.fill(HIST("event/trackselection"), 2.5f);
-      if (!t.hasTOF()) { // Skipping tracks without TOF
-        continue;
-      }
-      histos.fill(HIST("event/trackselection"), 3.5f);
-
-      const float tof = t.tofSignal() - collisionTime_ps;
-      histos.fill(HIST("event/particlehypo"), t.pidForTracking());
-      histos.fill(HIST("event/tofsignal"), t.p(), t.tofSignal());
-      histos.fill(HIST("event/pexp"), t.p(), t.tofExpMom());
-      histos.fill(HIST("event/eta"), t.eta());
-      histos.fill(HIST("event/phi"), t.phi());
-      histos.fill(HIST("event/etaphi"), t.eta(), t.phi());
-      histos.fill(HIST("event/length"), t.length());
-      histos.fill(HIST("event/pt"), t.pt());
-      histos.fill(HIST("event/p"), t.p());
-      // histos.fill(HIST("event/ptreso"), t.p(), t.sigma1Pt() * t.pt() * t.pt());
-      //
-      fillParticleHistos<PID::Electron>(t, tof);
-      fillParticleHistos<PID::Muon>(t, tof);
-      fillParticleHistos<PID::Pion>(t, tof);
-      fillParticleHistos<PID::Kaon>(t, tof);
-      fillParticleHistos<PID::Proton>(t, tof);
-      fillParticleHistos<PID::Deuteron>(t, tof);
-      fillParticleHistos<PID::Triton>(t, tof);
-      fillParticleHistos<PID::Helium3>(t, tof);
-      fillParticleHistos<PID::Alpha>(t, tof);
-    }
-  }
-};
-
 WorkflowSpec defineDataProcessing(ConfigContext const& cfgc)
 {
   auto workflow = WorkflowSpec{adaptAnalysisTask<tofPidFull>(cfgc)};
   if (cfgc.options().get<int>("add-qa")) {
-    workflow.push_back(adaptAnalysisTask<tofPidFullQa>(cfgc));
+    workflow.push_back(adaptAnalysisTask<tofPidQa>(cfgc));
   }
 
   return workflow;
